Use c_ubytes for storing binary data
Previously, c_uchar type was used to store binary data.
The Python documentation clearly states that this type should only be
used for immutable data, and for storing actual strings of characters
without null chars inside.

Switched to c_ubyte type so that mutable type is used for buffers
which are modified by kernel.